+++ app/api/v1/gmail/labels/route.ts
@@ -10,14 +10,14 @@ import { Label, Label_Basic } from '@/interfaces/gmail.interfaces';
 import { getGmailToken } from '@/lib/utils/nango/google.server';
 
 export async function GET(req: NextRequest): Promise<NextResponse> {
-  const userId = req.nextUrl.searchParams.get('userId');
+  const companyId = req.nextUrl.searchParams.get('companyId');
 
-  if (!userId) {
+  if (!companyId) {
     return badRequest('User ID is required');
   }
 
   try {
-    const googleMailToken = await getGmailToken(userId);
+    const googleMailToken = await getGmailToken(companyId);
 
     if (!googleMailToken) {
       return unauthorized('Google Mail token not found');
@@ -41,15 +41,17 @@ export async function GET(req: NextRequest): Promise<NextResponse> {
 }
 
 export async function POST(req: NextRequest): Promise<NextResponse> {
-  const { userId, label }: { userId: string; label: Omit<Label_Basic, 'id'> } =
-    await req.json();
+  const {
+    companyId,
+    label,
+  }: { companyId: string; label: Omit<Label_Basic, 'id'> } = await req.json();
 
-  if (!userId || !label) {
+  if (!companyId || !label) {
     return badRequest('User ID and label are required');
   }
 
   try {
-    const googleMailToken = await getGmailToken(userId);
+    const googleMailToken = await getGmailToken(companyId);
 
     if (!googleMailToken) {
       return unauthorized('Google Mail token not found');
